+++ NEWS.md
@@ -1,5 +1,14 @@
 
 <!-- NEWS.md is generated from NEWS.Rmd. Please edit that file -->
+splithalf v0.2.0 (current development version)
+==============================================
+
+Changes:
+
+-   Added the TSTsplithalf function
+-   Some minor bug fixes
+-   Output for all functions now includes two-part alpha
+
 splithalf v0.1.0 (Release data: 07/04/2017)
 ===========================================